requirements.txt
streamlit
supabase
python-dotenv
groq